requirements.txt
# Python dependencies for AI scoring module

# AI/ML libraries
google-generativeai==0.3.2
openai==1.6.1

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
